requirement.txt
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.2
alpha_vantage==2.3.1
scipy==1.12.0
requests==2.31.0
python-dotenv==1.0.0